Server/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==1.8.1
Flask==1.1.2
requests==2.24.0
functions-framework==2.0.0
flask_restful
flask_cors
dnspython==2.0.0
PyJWT==2.0.0
PyMySQL==0.10.1
Flask-MySQL==1.5.1

